+++ nanodjango/constants.py
@@ -1,0 +1,1 @@
+SQLITE_MEMORY = ":memory:"